+++ test/src/RecordWithManyFields.mjs
@@ -2,7 +2,6 @@
 
 import * as Spice from "./Spice.mjs";
 import * as Stdlib_Option from "@rescript/runtime/lib/es6/Stdlib_Option.js";
-import * as Stdlib_Result from "@rescript/runtime/lib/es6/Stdlib_Result.js";
 
 function params_encode(v) {
   let extra = v.width;
@@ -168,77 +167,77 @@ function params_decode(v) {
   if (typeof v !== "object" || v === null || Array.isArray(v)) {
     return Spice.error(undefined, "Not an object", v);
   }
-  let width = Stdlib_Option.getOr(Stdlib_Option.map(v["width"], json => Stdlib_Result.map(Spice.floatFromJson(json), v => v)), {
+  let width = Stdlib_Option.getOr(Stdlib_Option.map(v["width"], extra => Spice.optionalFieldFromJson(Spice.floatFromJson, extra)), {
     TAG: "Ok",
     _0: undefined
   });
   if (width.TAG === "Ok") {
-    let height = Stdlib_Option.getOr(Stdlib_Option.map(v["height"], json => Stdlib_Result.map(Spice.floatFromJson(json), v => v)), {
+    let height = Stdlib_Option.getOr(Stdlib_Option.map(v["height"], extra => Spice.optionalFieldFromJson(Spice.floatFromJson, extra)), {
       TAG: "Ok",
       _0: undefined
     });
     if (height.TAG === "Ok") {
-      let animation = Stdlib_Option.getOr(Stdlib_Option.map(v["animation"], json => Stdlib_Result.map(Spice.boolFromJson(json), v => v)), {
+      let animation = Stdlib_Option.getOr(Stdlib_Option.map(v["animation"], extra => Spice.optionalFieldFromJson(Spice.boolFromJson, extra)), {
         TAG: "Ok",
         _0: undefined
       });
       if (animation.TAG === "Ok") {
-        let focusInput = Stdlib_Option.getOr(Stdlib_Option.map(v["focusInput"], json => Stdlib_Result.map(Spice.boolFromJson(json), v => v)), {
+        let focusInput = Stdlib_Option.getOr(Stdlib_Option.map(v["focusInput"], extra => Spice.optionalFieldFromJson(Spice.boolFromJson, extra)), {
           TAG: "Ok",
           _0: undefined
         });
         if (focusInput.TAG === "Ok") {
-          let autoMapping = Stdlib_Option.getOr(Stdlib_Option.map(v["autoMapping"], json => Stdlib_Result.map(Spice.boolFromJson(json), v => v)), {
+          let autoMapping = Stdlib_Option.getOr(Stdlib_Option.map(v["autoMapping"], extra => Spice.optionalFieldFromJson(Spice.boolFromJson, extra)), {
             TAG: "Ok",
             _0: undefined
           });
           if (autoMapping.TAG === "Ok") {
-            let shorthand = Stdlib_Option.getOr(Stdlib_Option.map(v["shorthand"], json => Stdlib_Result.map(Spice.boolFromJson(json), v => v)), {
+            let shorthand = Stdlib_Option.getOr(Stdlib_Option.map(v["shorthand"], extra => Spice.optionalFieldFromJson(Spice.boolFromJson, extra)), {
               TAG: "Ok",
               _0: undefined
             });
             if (shorthand.TAG === "Ok") {
-              let pleaseReadGuide = Stdlib_Option.getOr(Stdlib_Option.map(v["pleaseReadGuide"], json => Stdlib_Result.map(Spice.intFromJson(json), v => v)), {
+              let pleaseReadGuide = Stdlib_Option.getOr(Stdlib_Option.map(v["pleaseReadGuide"], extra => Spice.optionalFieldFromJson(Spice.intFromJson, extra)), {
                 TAG: "Ok",
                 _0: undefined
               });
               if (pleaseReadGuide.TAG === "Ok") {
-                let pleaseReadGuideTimer = Stdlib_Option.getOr(Stdlib_Option.map(v["pleaseReadGuideTimer"], json => Stdlib_Result.map(Spice.floatFromJson(json), v => v)), {
+                let pleaseReadGuideTimer = Stdlib_Option.getOr(Stdlib_Option.map(v["pleaseReadGuideTimer"], extra => Spice.optionalFieldFromJson(Spice.floatFromJson, extra)), {
                   TAG: "Ok",
                   _0: undefined
                 });
                 if (pleaseReadGuideTimer.TAG === "Ok") {
-                  let maxSuggestItems = Stdlib_Option.getOr(Stdlib_Option.map(v["maxSuggestItems"], json => Stdlib_Result.map(Spice.intFromJson(json), v => v)), {
+                  let maxSuggestItems = Stdlib_Option.getOr(Stdlib_Option.map(v["maxSuggestItems"], extra => Spice.optionalFieldFromJson(Spice.intFromJson, extra)), {
                     TAG: "Ok",
                     _0: undefined
                   });
                   if (maxSuggestItems.TAG === "Ok") {
-                    let showMoreHName = Stdlib_Option.getOr(Stdlib_Option.map(v["showMoreHName"], json => Stdlib_Result.map(Spice.boolFromJson(json), v => v)), {
+                    let showMoreHName = Stdlib_Option.getOr(Stdlib_Option.map(v["showMoreHName"], extra => Spice.optionalFieldFromJson(Spice.boolFromJson, extra)), {
                       TAG: "Ok",
                       _0: undefined
                     });
                     if (showMoreHName.TAG === "Ok") {
-                      let hideMapBtn = Stdlib_Option.getOr(Stdlib_Option.map(v["hideMapBtn"], json => Stdlib_Result.map(Spice.boolFromJson(json), v => v)), {
+                      let hideMapBtn = Stdlib_Option.getOr(Stdlib_Option.map(v["hideMapBtn"], extra => Spice.optionalFieldFromJson(Spice.boolFromJson, extra)), {
                         TAG: "Ok",
                         _0: undefined
                       });
                       if (hideMapBtn.TAG === "Ok") {
-                        let hideEngBtn = Stdlib_Option.getOr(Stdlib_Option.map(v["hideEngBtn"], json => Stdlib_Result.map(Spice.boolFromJson(json), v => v)), {
+                        let hideEngBtn = Stdlib_Option.getOr(Stdlib_Option.map(v["hideEngBtn"], extra => Spice.optionalFieldFromJson(Spice.boolFromJson, extra)), {
                           TAG: "Ok",
                           _0: undefined
                         });
                         if (hideEngBtn.TAG === "Ok") {
-                          let alwaysShowEngAddr = Stdlib_Option.getOr(Stdlib_Option.map(v["alwaysShowEngAddr"], json => Stdlib_Result.map(Spice.boolFromJson(json), v => v)), {
+                          let alwaysShowEngAddr = Stdlib_Option.getOr(Stdlib_Option.map(v["alwaysShowEngAddr"], extra => Spice.optionalFieldFromJson(Spice.boolFromJson, extra)), {
                             TAG: "Ok",
                             _0: undefined
                           });
                           if (alwaysShowEngAddr.TAG === "Ok") {
-                            let useBannerLink = Stdlib_Option.getOr(Stdlib_Option.map(v["useBannerLink"], json => Stdlib_Result.map(Spice.boolFromJson(json), v => v)), {
+                            let useBannerLink = Stdlib_Option.getOr(Stdlib_Option.map(v["useBannerLink"], extra => Spice.optionalFieldFromJson(Spice.boolFromJson, extra)), {
                               TAG: "Ok",
                               _0: undefined
                             });
                             if (useBannerLink.TAG === "Ok") {
-                              let submitMode = Stdlib_Option.getOr(Stdlib_Option.map(v["submitMode"], json => Stdlib_Result.map(Spice.boolFromJson(json), v => v)), {
+                              let submitMode = Stdlib_Option.getOr(Stdlib_Option.map(v["submitMode"], extra => Spice.optionalFieldFromJson(Spice.boolFromJson, extra)), {
                                 TAG: "Ok",
                                 _0: undefined
                               });